backend/requirements.txt
pyserial
Flask
flask
flask-cors
requests
pygame
python-dotenv
openai
